Correction and improvements App for the Lesson 8
