+++ module_L6/src/main/java/org/example/controllers/ProductServiceController.java
@@ -3,10 +3,13 @@ package org.example.controllers;
 import org.example.entity.Customer;
 import org.example.entity.Product;
 import org.example.services.*;
+import org.example.services.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import java.math.BigDecimal;
@@ -17,23 +20,23 @@ import java.util.Optional;
 public class ProductServiceController {
 
     @Autowired
-    private EntityService<Product> productService;
+    private ProductService<Product> productServiceImpl;
     @Autowired
-    private EntityService<Customer> customerService;
+    private CustomerService<Customer> customerServiceImpl;
     @Autowired
-    private OrdersService<Customer, Product> ordersService;
+    private OrdersService<Customer, Product> ordersServiceImpl;
 
     private Customer customer;
 
     @GetMapping("/index")
     public String auth(HttpServletRequest request) {
-        customer = customerService.getByName(request.getParameter("name"));
+        customer = customerServiceImpl.getByName(request.getParameter("name"));
         return "redirect:product_list";
     }
 
     @PostMapping("/form_product")
     public String formUpdateProduct(Product product) {
-        productService.save(product);
+        productServiceImpl.save(product);
         return "redirect:product_list";
     }
 
@@ -43,9 +46,13 @@ public class ProductServiceController {
                                   @RequestParam(name = "max", required = false)Optional<BigDecimal> max,
                                   @RequestParam(name = "page", required = false)Optional<Integer> page,
                                   @RequestParam(name = "size", required = false)Optional<Integer> size) {
-        model.addAttribute("products", productService.getByThroughFilter(min, max, page, size));
-        model.addAttribute("name", customer.getName());
-        model.addAttribute("customer_products", ordersService.getListProducts(customer.getId()));
+        try {
+            model.addAttribute("products", productServiceImpl.getByThroughFilter(min, max, page, size));
+            model.addAttribute("name", customer.getName());
+            model.addAttribute("customer_products", ordersServiceImpl.getListProducts(customer.getId()));
+        }catch (RuntimeException e){
+            throw new NotFoundException();
+        }
         return "product_list";
     }
 
@@ -57,43 +64,45 @@ public class ProductServiceController {
 
     @GetMapping("{id}")
     public String updateProduct(@PathVariable(value = "id") Long id, Model model) {
-        model.addAttribute("product", productService.getById(id));
+        model.addAttribute("product", productServiceImpl.getById(id));
         return "form_product";
     }
 
     @GetMapping("/delete/{id}")
     public String deleteProduct(@PathVariable(value = "id") Long id) {
-        productService.remove(id);
+        productServiceImpl.remove(id);
         return "redirect:/index/product_list";
     }
 
     @GetMapping("/buy/{id}")
     public String buyProduct(@PathVariable(value = "id") Long id) {
-        ordersService.addProductToCustomer(customer, productService.getById(id));
+        ordersServiceImpl.addProductToCustomer(customer, productServiceImpl.getById(id));
         return "redirect:/index/product_list";
     }
 
     @GetMapping("/delete_purcha/{id}")
     public String deletePurcha(@PathVariable(value = "id") Long id) {
-        ordersService.delProductToCustomer(customer, productService.getById(id));
+        ordersServiceImpl.delProductToCustomer(customer, productServiceImpl.getById(id));
         return "redirect:/index/product_list";
     }
 
     @GetMapping("/buyer_list/{id}")
     public String showBuyerList(@PathVariable(value = "id") Long id, Model model) {
-        model.addAttribute("name_product", productService.getById(id).getTitle());
-        model.addAttribute("customers", ordersService.getListCustomers(id));
+        model.addAttribute("name_product", productServiceImpl.getById(id).getTitle());
+        model.addAttribute("customers", ordersServiceImpl.getListCustomers(id));
         return "buyer_list";
     }
 
     @GetMapping("/orders_list")
     public String showOrderList(Model model) {
-        model.addAttribute("orders", ordersService.getAllOrders());
+        model.addAttribute("orders", ordersServiceImpl.getAllOrders());
         return "orders_list";
     }
 
-    @GetMapping("/404")
-    public String showErrorFindUser() {
-        return "404";
+    @ExceptionHandler
+    public ModelAndView notFoundExceptionHandler(NotFoundException e) {
+        ModelAndView modelAndView = new ModelAndView("/404");
+        modelAndView.setStatus(HttpStatus.NOT_FOUND);
+        return modelAndView;
     }
 }
