+++ module_L6/src/main/java/org/example/services/CustomerService.java
@@ -1,53 +1,18 @@
 package org.example.services;
 
-import lombok.AllArgsConstructor;
 import org.example.entity.Customer;
-import org.example.repository.CustomerRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.List;
-import java.util.Optional;
 
-@Service
-@AllArgsConstructor
-public class CustomerService implements EntityService<Customer> {
+public interface CustomerService<T> {
 
-    private CustomerRepository customerRepository;
+    Customer getById(Long id);
 
-    @Override
-    @Transactional(readOnly = true)
-    public Customer getById(Long id) {
-        return customerRepository.findById(id).get();
-    }
+    void save(T t);
 
-    @Override
-    @Transactional
-    public void save(Customer customer) {
-        customerRepository.save(customer);
-    }
+    List<Customer> getEntityAll();
 
-    @Override
-    @Transactional(readOnly = true)
-    public List<Customer> getEntityAll() {
-        return (List<Customer>) customerRepository.findAll();
-    }
+    void remove(Long id);
 
-    @Override
-    @Transactional
-    public void remove(Long id) {
-        customerRepository.deleteById(id);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public Customer getByName(String name) {
-        return customerRepository.findByName(name);
-    }
-
-    @Override
-    public List<Customer> getByThroughFilter(Optional<BigDecimal> param1, Optional<BigDecimal> param2) {
-        return null;
-    }
+    Customer getByName(String name);
 }
