+++ module_L6/src/main/java/org/example/services/ProductService.java
@@ -1,66 +1,22 @@
 package org.example.services;
 
-import lombok.AllArgsConstructor;
-import org.example.entity.Product;
-import org.example.repository.ProductRepository;
-import org.example.repository.specifications.ProductSpecification;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
-@Service
-@AllArgsConstructor
-public class ProductService implements EntityService<Product>{
+public interface ProductService<T> {
 
-    private ProductRepository productRepository;
+    T getById(Long id);
 
-    @Override
-    @Transactional(readOnly = true)
-    public Product getById(Long id) {
-       return productRepository.findById(id).get();
-    }
+    void save(T t);
 
-    @Override
-    @Transactional
-    public void save(Product product) {
-        productRepository.save(product);
-    }
+    List<T> getEntityAll();
 
-    @Override
-    @Transactional(readOnly = true)
-    public List<Product> getEntityAll() {
-        return (List<Product>) productRepository.findAll();
-    }
+    void remove(Long id);
 
-    @Override
-    @Transactional
-    public void remove(Long id) {
-        productRepository.deleteById(id);
-    }
+    T getByName(String name);
 
-    @Override
-    @Transactional(readOnly = true)
-    public Product getByName(String title) {
-       return productRepository.findByTitle(title);
-    }
-
-    @Transactional(readOnly = true)
-    public Page<Product> getByThroughFilter(Optional<BigDecimal> min, Optional<BigDecimal> max, Optional<Integer> page, Optional<Integer> size) {
-
-        Specification<Product> specification = Specification.where(null);
-        if (min.isPresent()) {
-            specification = specification.and(ProductSpecification.greatOrEquals(min.get()));
-        }
-        if (max.isPresent()) {
-            specification = specification.and(ProductSpecification.lessOrEquals(max.get()));
-        }
-
-        return productRepository.findAll(specification, PageRequest.of(page.orElse(1) - 1, size.orElse(10)));
-    }
+    Page<T> getByThroughFilter(Optional<BigDecimal> min, Optional<BigDecimal> max, Optional<Integer> page, Optional<Integer> size);
 }
