+++ module_L6/src/main/java/org/example/services/exceptions/NotFoundException.java
@@ -1,0 +1,4 @@
+package org.example.services.exceptions;
+
+public class NotFoundException extends RuntimeException{
+}
