+++ module_L6/src/main/resources/templates/404.html
@@ -1,11 +1,27 @@
-<!DOCTYPE html>
+<!doctype html>
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
-    <meta charset="UTF-8">
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
     <title>404</title>
 </head>
-<body>
-<h2 align="center">This user does not exist</h2>
-<a th:href="/">Back</a>
+
+<div class="container">
+    <section class="errot-container text-center">
+        <h1>404</h1>
+        <div class="error-divider">
+            <h2>NOT FOUND</h2>
+        </div>
+        <a href="/" class="return-btn">Return</a>
+    </section>
+</div>
+
+
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
+        crossorigin="anonymous"></script>
+<script src="https://kit.fontawesome.com/944a576475.js" crossorigin="anonymous"></script>
 </body>
 </html>
