+++ module_L6/src/main/resources/templates/index.html
@@ -11,9 +11,6 @@
         <button type="submit" class="btn btn-primary">View all products</button>
     </form>
 </div>
-<!--    <a href="/index/add_product">Add product</a>-->
-<!--<br>-->
-<!--<br>-->
-<!--    <a href="/index/product_list">View all products</a>-->
+
 </body>
 </html>
