+++ module_L6/src/main/resources/templates/product_list.html
@@ -8,86 +8,96 @@
     <title>Product list</title>
 </head>
 <body>
-<h1 th:align="center" th:text="'Welcome, ' + ${name}"/>
+
 <div class="container">
-    <div class="col-10 mx-auto">
-        <h2>Product list</h2>
+    <h1 th:align="center" th:text="'Welcome, ' + ${name}"/>
+    <div>
+        <div class="col-10 mx-auto">
+            <h2>Product list</h2>
 
-        <div class="col-9">
-            <form class="form-inline" th:method="get">
-                <input type="number" th:value="${param.min}" class="form-control-md" name="min" placeholder="Min">
-                <input type="number" th:value="${param.max}" class="form-control-md" name="max" placeholder="Max">
-                <button type="submit" class="btn btn-outline-dark" th:text="Filter"></button>
-            </form>
+            <div class="col-9">
+                <form class="form-inline" th:method="get">
+                    <input type="number" th:value="${param.min}" class="form-control-md" name="min" placeholder="Min">
+                    <input type="number" th:value="${param.max}" class="form-control-md" name="max" placeholder="Max">
+                    <button type="submit" class="btn btn-outline-dark" th:text="Filter"></button>
+                </form>
+            </div>
+            <table class="table table-success table-striped">
+                <thead>
+                <tr>
+                    <th scope="col">id</th>
+                    <th scope="col">title</th>
+                    <th scope="col">cost</th>
+                </tr>
+                </thead>
+                <tbody>
+                <tr th:if="${products.hasContent()}" th:each="product : ${products}">
+                    <th scope="row" th:text="${product.id}">1</th>
+                    <th scope="row" th:text="${product.title}">1</th>
+                    <th scope="row" th:text="${product.cost}">1</th>
+                    <th>
+                        <a th:href="@{/index/buy/{id}(id = ${product.id})}">Buy</a>
+                        <a th:href="@{/index/{id}(id = ${product.id})}">Update</a>
+                        <a th:href="@{/index/delete/{id}(id = ${product.id})}">Delete</a>
+                        <a th:href="@{/index/buyer_list/{id}(id = ${product.id})}">Show buyer List</a>
+                    </th>
+                </tr>
+
+                <tr th:unless="${products.hasContent()}">
+                    <td colspan="3">No products</td>
+                </tr>
+                </tbody>
+            </table>
+            <div>
+                <a th:href="@{/index/add_product}">Add product</a>
+                <a th:href="@{/index/orders_list}">View order list</a>
+            </div>
         </div>
-        <table class="table table-success table-striped">
-            <thead>
-            <tr>
-                <th scope="col">id</th>
-                <th scope="col">title</th>
-                <th scope="col">cost</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:if="${products.hasContent()}" th:each="product : ${products}">
-                <th scope="row" th:text="${product.id}">1</th>
-                <th scope="row" th:text="${product.title}">1</th>
-                <th scope="row" th:text="${product.cost}">1</th>
-                <th>
-                    <a th:href="@{/index/buy/{id}(id = ${product.id})}">Buy</a>
-                    <a th:href="@{/index/{id}(id = ${product.id})}">Update</a>
-                    <a th:href="@{/index/delete/{id}(id = ${product.id})}">Delete</a>
-                    <a th:href="@{/index/buyer_list/{id}(id = ${product.id})}">Show buyer List</a>
-                </th>
-            </tr>
-
-            <tr th:unless="${products.hasContent()}">
-                <td colspan="3">No products</td>
-            </tr>
-            </tbody>
-        </table>
-        <div>
-            <a th:href="@{/index/add_product}">Add product</a>
-            <a th:href="@{/index/orders_list}">View order list</a>
+        <div th:if="${products.hasContent()}">
+            <nav>
+                <ul class="pagination justify-content-center">
+                    <li>
+                        <a class="page-link" th:if="${products.getPageable().getPageNumber() > 0}" th:href="@{/index/product_list(page=${products.getPageable().getPageNumber()}, size=${products.size}, min=${param.min}, max=${param.max})}">Previous</a>
+                    </li>
+                    <li th:each="number : ${#numbers.sequence(1, products.totalPages)}"
+                        class="page-item" th:classappend="${number == products.number + 1} ? 'active'">
+                        <a class="page-link" th:href="@{/index/product_list(page=${number}, size=${products.size}, min=${param.min}, max=${param.max})}">[[${number}]]</a>
+                    </li>
+                    <li>
+                        <a class="page-link" th:if="${products.hasNext()}" th:href="@{/index/product_list(page=${products.nextPageable().getPageNumber() + 1}, size=${products.size}, min=${param.min}, max=${param.max})}">Next</a>
+                    </li>
+                </ul>
+            </nav>
         </div>
     </div>
-    <div th:if="${products.hasContent()}">
-        <nav>
-            <ul class="pagination justify-content-center">
-                <li th:each="number : ${#numbers.sequence(1, products.totalPages)}"
-                    class="page-item" th:classappend="${number == products.number + 1} ? 'active'">
-                    <a class="page-link" th:href="@{/index/product_list(page=${number}, size=${products.size}, min=${param.min}, max=${param.max})}">[[${number}]]</a>
-                </li>
-            </ul>
-        </nav>
+    <br>
+    <br>
+    <div>
+        <div class="col-10 mx-auto">
+            <h2>list of purchased</h2>
+            <table class="table table-success table-striped">
+                <thead>
+                <tr>
+                    <th scope="col">id</th>
+                    <th scope="col">title</th>
+                    <th scope="col">cost</th>
+                </tr>
+                </thead>
+                <tbody>
+                <tr th:each="product : ${customer_products}">
+                    <th scope="row" th:text="${product.id}">1</th>
+                    <th scope="row" th:text="${product.title}">1</th>
+                    <th scope="row" th:text="${product.cost}">1</th>
+                    <th>
+                        <a th:href="@{/index/delete_purcha/{id}(id = ${product.id})}">Delete</a>
+                    </th>
+                </tr>
+                </tbody>
+            </table>
+        </div>
     </div>
 </div>
-<br>
-<br>
-<div class="container">
-    <div class="col-10 mx-auto">
-        <h2>list of purchased</h2>
-        <table class="table table-success table-striped">
-            <thead>
-            <tr>
-                <th scope="col">id</th>
-                <th scope="col">title</th>
-                <th scope="col">cost</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:each="product : ${customer_products}">
-                <th scope="row" th:text="${product.id}">1</th>
-                <th scope="row" th:text="${product.title}">1</th>
-                <th scope="row" th:text="${product.cost}">1</th>
-                <th>
-                    <a th:href="@{/index/delete_purcha/{id}(id = ${product.id})}">Delete</a>
-                </th>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-</div>
+
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
         crossorigin="anonymous"></script>
